guacamole: overlay sensor names for OnePlusDoze
As suggested by eyosen, we overlay the sensor names to make the vendor 
name match with the name in our OnePlusDoze package. This way keeping 
the package uniform for all devices and make the options in the doze 
screen visible.

note: this can be done via fwb overlays too, but the options will then 
be displayed outside the Doze config screen and in the main settings.