guacamole: just leave the dependency for firmware as info
The repo is only needed on the initial build AND to make sure the user is on the appropriate vendor version on flashing.
Otherwise it only make the build larger and constant flashing of manufacturer partitions is not need, except we update the vendor.
(Unlikely happen very often as the device has reached its last stage of support)